requirements.txt
pandas
numpy
matplotlib.pyplot
seaborn
warnings